refactor: move parsing issue methods to parsing issue service
